Adjusted Path Logic to Support Linux
Ensured that relative filepaths are resolved correctly against the working directory, rather than as absolute paths.
Also implemented a new flag "-r" which attempts to repair incorrect capitalization in filepaths! A lot of builds use codeset files with capitalization mistakes in their .include directives that don't affect Windows (because the filesystems there are case-insensitive), but *do* break Linux. This should allow the builder to identify those files in spite of those mistakes, though fixing the errors would be preferable.